Refactor Conductor for middleware components
This is a major refactor of Conductor. It is now only for creating a sequence of middleware components. All routing capabilities as well as stock middleware components are now removed. These may come back as separate packages.

The goal of the project is to make it easy to define a sequence of middleware using a standard pattern of wrapper http.Handler:

```go
func(http.Handler) http.Handler
```